requirements.txt
llama-cpp-python==0.2.79
fastapi==0.111.0
pydantic==2.7.4
uvicorn==0.30.1